Add correct reading of keymap
Previous commits used UART to send the USB HID keycodes from the secondary
half to the primary half. The problem was that I forgot there could be
layer key combinations with one half pressing the layer key and the other
pressing a normal key. Therefore, I changed the communication to send the
pressed key indices of the secondary half. The changed code improves the
polling rate of the primary half to 70kHz and 80kHz on the secondary half.
My test setup consists of just two Picos with a switch on each, so I cannot
test the typing feel or if fast typing works well.